Review

+++ backend/src/models/product.ts
@@ -29,7 +29,13 @@ const cardsSchema = new mongoose.Schema<IProduct>(
                 type: String,
                 required: [true, 'Поле "image.fileName" должно быть заполнено'],
             },
-            originalName: String,
+            originalName: {
+                type: String,
+                required: [
+                    true,
+                    'Поле "image.originalName" должно быть заполнено',
+                ],
+            },
         },
         category: {
             type: String,
@@ -50,10 +56,15 @@ cardsSchema.index({ title: 'text' })
 
 // Можно лучше: удалять старое изображением перед обновлением сущности
 cardsSchema.pre('findOneAndUpdate', async function deleteOldImage() {
-    // @ts-ignore
+    // @ts-expect-error - Mongoose types don't properly type this.getUpdate()
     const updateImage = this.getUpdate().$set?.image
     const docToUpdate = await this.model.findOne(this.getQuery())
-    if (updateImage && docToUpdate) {
+    if (
+        updateImage &&
+        docToUpdate &&
+        docToUpdate.image &&
+        docToUpdate.image.fileName
+    ) {
         unlink(
             join(__dirname, `../public/${docToUpdate.image.fileName}`),
             (err) => console.log(err)
@@ -63,9 +74,13 @@ cardsSchema.pre('findOneAndUpdate', async function deleteOldImage() {
 
 // Можно лучше: удалять файл с изображением после удаление сущности
 cardsSchema.post('findOneAndDelete', async (doc: IProduct) => {
-    unlink(join(__dirname, `../public/${doc.image.fileName}`), (err) =>
-        console.log(err)
-    )
+    if (doc.image && doc.image.fileName) {
+        unlink(join(__dirname, `../public/${doc.image.fileName}`), (err) =>
+            console.log(err)
+        )
+    }
 })
 
-export default mongoose.model<IProduct>('product', cardsSchema)
+const ProductModel = mongoose.model<IProduct>('products', cardsSchema)
+console.log('Product model collection name:', ProductModel.collection.name)
+export default ProductModel
